renderer: Remove render_texture from Frame
